# api/openapi.yaml
openapi: 3.1.0
info:
  title: Hubtask API
  version: 0.1.0
  summary: Task management with Hub / Collection / Task / Work Package / Activity
  description: |
    The skeleton of the API contract (API first). This file is the single source of truth:
    server interfaces and client SDKs are generated from it (`make generate`).

    For the conventions see docs/architecture/api-guidelines.md:
      * one major path `/api/v1`, additive changes without a version bump
      * errors per RFC 9457 with a stable `code`
      * cursor pagination, `Idempotency-Key`, `ETag`/`If-Match`
      * actions as a suffix (`POST /items/{id}:complete`)
      * no display text, only codes and parameters

    Not yet elaborated (deliberately open, to be added in phase 0/1): templates, views, jumble,
    reminders, recurrence, automation, webhooks, media, admin/tenants, search, SSE.
  license:
    name: BUSL-1.1
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /containers:
    get:
      tags: [containers]
      operationId: listContainers
      parameters:
        - $ref: "#/components/parameters/ContainerTypeFilter"
        - $ref: "#/components/parameters/ParentId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/IncludeArchived"
      responses:
        "200":
          description: A page of containers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerPage"
        "4XX":
          $ref: "#/components/responses/Problem"
    post:
      tags: [containers]
      operationId: createContainer
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerCreate"
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
        "4XX":
          $ref: "#/components/responses/Problem"
...
  /containers/{containerId}/buckets:
    parameters:
      - $ref: "#/components/parameters/ContainerId"
    get:
      tags: [structure]
      operationId: listBuckets
      responses:
        "200":
          description: The collection's buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bucket"
        "4XX":
          $ref: "#/components/responses/Problem"
    post:
      tags: [structure]
      operationId: createBucket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BucketCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bucket"
        "4XX":
          $ref: "#/components/responses/Problem"
  /containers/{containerId}/buckets/{bucketId}:
    parameters:
      - $ref: "#/components/parameters/ContainerId"
      - $ref: "#/components/parameters/BucketId"
    patch:
      tags: [structure]
      operationId: updateBucket
      description: >
        Changes a column's own fields. A field that is not sent is left alone; sending color_token
        as null clears it, and sending wip_limit as null or 0 removes the limit. The name stays
        unique in the collection. Where the column sits is not here - that is :reorder, so that a
        rename and a rearrangement are two operations with two audit entries.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/BucketUpdate"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bucket"
        "4XX":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [structure]
      operationId: deleteBucket
      description: >
        Takes the column off the board. Its entries are not deleted with it: they move to the
        leftmost remaining column, or out of any column when this was the last one. What became of
        them is in the answer rather than left to be discovered.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Deleted, with what became of the entries that were in it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BucketDeletion"
        "4XX":
          $ref: "#/components/responses/Problem"
  /containers/{containerId}/buckets/{bucketId}:reorder:
    parameters:
      - $ref: "#/components/parameters/ContainerId"
      - $ref: "#/components/parameters/BucketId"
    post:
      tags: [structure]
      operationId: reorderBucket
      description: >
        Moves a column along its board. An absent before_bucket_id moves it to the right hand end.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                before_bucket_id:
                  type: [string, "null"]
                  format: uuid
                  description: The column this one goes to the left of. Absent is the right hand end.
      responses:
        "200":
          description: Moved
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bucket"
        "4XX":
          $ref: "#/components/responses/Problem"
  /containers/{containerId}/labels:
    parameters:
      - $ref: "#/components/parameters/ContainerId"
    get:
      tags: [structure]
      operationId: listLabels
      responses:
        "200":
          description: The collection's labels, by name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
        "4XX":
          $ref: "#/components/responses/Problem"
    post:
      tags: [structure]
      operationId: createLabel
      description: >
        Adds a label to a collection's vocabulary. The name has to be free in the collection,
        compared without regard to case or accents. The colour is required: a label is rendered as
        a chip and nothing else, so with none a client would have to invent one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelCreate"
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "4XX":
          $ref: "#/components/responses/Problem"
  /containers/{containerId}/labels/{labelId}:
    parameters:
      - $ref: "#/components/parameters/ContainerId"
      - $ref: "#/components/parameters/LabelId"
    patch:
      tags: [structure]
      operationId: updateLabel
      description: >
        Changes a label's own fields. A field that is not sent is left alone; sending description as
        null clears it. The colour cannot be cleared - a label is rendered as a chip and nothing
        else. The name stays unique in the collection.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/LabelUpdate"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Label"
        "4XX":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [structure]
      operationId: deleteLabel
      description: >
        Takes the label out of the collection's vocabulary. The entries that carried it are not
        rewritten: they stop showing it because it is gone, which is what makes the deletion
        undoable without having to remember who wore it.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted
        "4XX":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "An OIDC access token, hbt_pat_…, or a service account token"
  parameters:
...
    ContainerId:
      name: containerId
      in: path
      required: true
      schema: { type: string, format: uuid }
    LabelId:
      name: labelId
      in: path
      required: true
      schema: { type: string, format: uuid }
    BucketId:
      name: bucketId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ParentId:
      name: parent_id
      in: query
      schema: { type: string, format: uuid }
    ContainerTypeFilter:
      name: type
      in: query
      schema:
        $ref: "#/components/schemas/ContainerType"
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    PageSize:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IncludeArchived:
      name: include_archived
      in: query
      schema: { type: boolean, default: false }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: A UUID; identical requests return the same result for 24 h.
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: The ETag of the state last read (optimistic locking).
      schema: { type: string }
  responses:
    Problem:
      description: An error per RFC 9457
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: A stable, machine-readable error code (part of the contract).
          examples: [capability_not_supported, version_conflict, rate_limited]
        detail_code:
          type: string
          description: The message code for localisation in the client.
        params:
          type: object
          additionalProperties: true
        field_errors:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              code: { type: string }
              params:
                type: object
                additionalProperties: true
        request_id: { type: string }
        docs: { type: string, format: uri }
    PageInfo:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, "null"]
          description: >
            The cursor for the next page, null on the last one. Opaque and signed: it is produced by
            the server and is neither to be constructed nor parsed by a client.
        has_more: { type: boolean }
    ContainerPage:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Container"
        page:
          $ref: "#/components/schemas/PageInfo"
...
    ContainerType:
      type: string
      enum: [HUB, COLLECTION]
...
    Container:
      type: object
      required: [id, type, name, version]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/ContainerType" }
        parent_id: { type: [string, "null"], format: uuid }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"] }
        icon: { type: [string, "null"] }
        color_token: { type: [string, "null"] }
        order_key: { type: string }
        policies:
          $ref: "#/components/schemas/ContainerPolicies"
        archived_at: { type: [string, "null"], format: date-time }
        effective_archived:
          type: boolean
          description: >
            Whether this container is read-only: archived itself, or sitting in an archived hub
            (I-C3). archived_at says which of the two - it is null for a collection that only
            inherits it.
        deleted_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
    ContainerCreate:
      type: object
      required: [type, name]
      properties:
        type: { $ref: "#/components/schemas/ContainerType" }
        parent_id: { type: [string, "null"], format: uuid }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, "null"] }
        icon: { type: [string, "null"] }
        color_token: { type: [string, "null"] }
...
    ContainerPolicies:
      type: object
      description: >
        How a collection works, as opposed to what it is called. One key today; the others the data
        model names - the default bucket, capability overrides, automatic assignment - arrive with
        the use cases that own them, and a key nothing reads would be a promise nothing keeps.
      properties:
        completion_policy: { $ref: "#/components/schemas/CompletionPolicy" }
    CompletionPolicy:
      type: string
      description: >
        Whether a child's completion propagates upwards. MANUAL leaves the parent alone; ROLLUP
        completes it when its last open child is completed and reopens it when any child is
        reopened. /meta/capabilities returns the valid values.
      enum: [MANUAL, ROLLUP]
      default: MANUAL
...
    Bucket:
      type: object
      description: >
        A column of a collection's board - the "list" of the requirements. It belongs to exactly one
        collection, and its name is unique there, compared without regard to case or accents.
      required: [id, collection_id, name, order_key, wip_limit, is_done_bucket, color_token, version]
      properties:
        id: { type: string, format: uuid }
        collection_id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        order_key:
          type: string
          description: >
            The rank on the board - a fractional index rather than a number (offline-sync.md §4.2).
        wip_limit:
          type: [integer, "null"]
          minimum: 1
          description: >
            How many entries the people using this board agreed to have in the column at once.
            Advisory: nothing refuses the drop that exceeds it, because the requirement is that the
            column turns red rather than that the work becomes impossible. Null is no limit.
        is_done_bucket: { type: boolean, default: false }
        color_token:
          type: [string, "null"]
          description: >
            Always present, as null when unset - like wip_limit. A board renders both, and a field
            that appeared only once somebody had set it is one a client cannot read unconditionally.
        version:
          type: integer
          description: The optimistic lock, returned as the ETag and sent back as If-Match.
    BucketCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        before_bucket_id: { type: [string, "null"], format: uuid }
        wip_limit: { type: [integer, "null"] }
        is_done_bucket: { type: boolean, default: false }
        color_token: { type: [string, "null"] }
    BucketUpdate:
      type: object
      description: JSON Merge Patch; null deletes a field.
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        wip_limit:
          type: [integer, "null"]
          minimum: 0
          description: >
            Null and 0 both remove the limit. Zero is not a limit anybody could satisfy, so nothing
            is lost by reading it that way, and it saves a client from a second field beside the
            number.
        is_done_bucket: { type: boolean }
        color_token: { type: [string, "null"] }
    BucketDeletion:
      type: object
      description: >
        What became of the entries that were in a deleted column. Reported rather than left to be
        discovered: a board that silently reassigned a person's cards would be indistinguishable
        from one that lost them.
      required: [bucket_id, target_bucket_id, moved_items]
      properties:
        bucket_id: { type: string, format: uuid }
        target_bucket_id:
          type: [string, "null"]
          format: uuid
          description: >
            The column the entries moved to - the leftmost remaining one. Null when this was the
            last column on the board, in which case the entries now carry none.
        moved_items:
          type: integer
          description: How many entries were moved.
    Label:
      type: object
      description: >
        A tag a collection defines and its entries carry. Defined on the collection rather than on
        the workspace: a label is a vocabulary the people working in one collection agree on, and a
        workspace-wide list would make every collection pay for every other's.
      required: [id, collection_id, name, color_token, description, version]
      properties:
        id: { type: string, format: uuid }
        collection_id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        color_token:
          type: string
          minLength: 1
          description: >
            A theme token rather than a colour value, so clients render it in their own palette.
            Required: a label is rendered as a chip and nothing else.
        description:
          type: [string, "null"]
          description: Always present, as null when unset.
        version:
          type: integer
          description: The optimistic lock, returned as the ETag and sent back as If-Match.
    LabelCreate:
      type: object
      required: [name, color_token]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        color_token: { type: string, minLength: 1 }
        description: { type: [string, "null"] }
    LabelUpdate:
      type: object
      description: JSON Merge Patch; null deletes a field.
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        color_token:
          type: string
          minLength: 1
          description: >
            Cannot be cleared. A label is rendered as a chip and nothing else, so with no colour a
            client would have to invent one.
        description: { type: [string, "null"] }
...